requirements.txt
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
PyMuPDF==1.23.8
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
requests==2.31.0
sentence-transformers>=2.2.2
huggingface_hub==0.17.3
transformers==4.34.1
pytest==7.4.0